Refactor delivery manager and add comments
