+++ typescript/food-ordering/app/src/restate-app/app.ts
@@ -1,7 +1,7 @@
 import * as restate from "@restatedev/restate-sdk";
 import * as driver from "./driver";
 import * as driver_pool from "./driver_pool";
-import * as delivery from "./delivery";
+import * as deliveryManager from "./delivery-manager";
 import * as orders from "./order_service";
 import * as orderstatus from "./order_status_service";
 import * as driverMobileAppSimulator from "./external/driver_mobile_app_sim";
@@ -13,7 +13,7 @@ if (require.main === module) {
       .bindKeyedRouter(orderstatus.service.path, orderstatus.router)
       .bindKeyedRouter(driver.service.path, driver.router)
       .bindKeyedRouter(driver_pool.service.path, driver_pool.router)
-      .bindKeyedRouter(delivery.service.path, delivery.router)
+      .bindKeyedRouter(deliveryManager.service.path, deliveryManager.router)
       .bindKeyedRouter(driverMobileAppSimulator.service.path, driverMobileAppSimulator.router)
       .listen(9080);
 }
